requirements.txt
python-telegram-bot[job-queue]>=21.0
python-dotenv>=1.0.0
tzdata>=2024.1
tweepy>=4.14,<5
